Split the CollectedLogger into two classes.
