+++ cli/src/main/java/de/jplag/cli/logger/CollectedLogger.java
@@ -2,7 +2,6 @@ package de.jplag.cli.logger;
 
 import org.slf4j.Marker;
 import org.slf4j.event.Level;
-import org.slf4j.helpers.AbstractLogger;
 import org.slf4j.helpers.MessageFormatter;
 
 import java.io.PrintStream;
@@ -11,20 +10,10 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.concurrent.ConcurrentLinkedDeque;
 
-public class CollectedLogger extends AbstractLogger {
-    private static final Level LOG_LEVEL_TRACE = Level.TRACE;
-    private static final Level LOG_LEVEL_DEBUG = Level.DEBUG;
-    private static final Level LOG_LEVEL_INFO = Level.INFO;
-    private static final Level LOG_LEVEL_WARN = Level.WARN;
-    private static final Level LOG_LEVEL_ERROR = Level.ERROR;
-
-    /**
-     * The default log level that shall be used for external libraries (like Stanford Core NLP)
-     */
-    private static final Level LOG_LEVEL_FOR_EXTERNAL_LIBRARIES = LOG_LEVEL_ERROR;
-
-    private static final Level CURRENT_LOG_LEVEL = LOG_LEVEL_INFO;
-
+/**
+ * A logger implementation, that prints all errors during finalization
+ */
+public class CollectedLogger extends JPlagLoggerBase {
     private static final int MAXIMUM_MESSAGE_LENGTH = 32;
 
     private static final PrintStream TARGET_STREAM = System.out;
@@ -41,7 +30,7 @@ public class CollectedLogger extends AbstractLogger {
     private final ConcurrentLinkedDeque<LogEntry> allErrors = new ConcurrentLinkedDeque<>();
 
     public CollectedLogger(String name) {
-        this.name = name;
+        super(LOG_LEVEL_INFO, name);
     }
 
     @Override
@@ -87,7 +76,7 @@ public class CollectedLogger extends AbstractLogger {
         this.isFinalizing = true;
         ArrayList<LogEntry> errors = new ArrayList<>(this.allErrors);
 
-        if(!errors.isEmpty()) {
+        if (!errors.isEmpty()) {
             info("Summary of all errors:");
             this.allErrors.removeAll(errors);
             for (LogEntry errorEntry : errors) {
@@ -98,70 +87,7 @@ public class CollectedLogger extends AbstractLogger {
         this.isFinalizing = false;
     }
 
-    @Override
-    public boolean isTraceEnabled() {
-        return isLogLevelEnabled(LOG_LEVEL_TRACE);
-    }
-
-    @Override
-    public boolean isTraceEnabled(Marker marker) {
-        return isTraceEnabled();
-    }
-
-    @Override
-    public boolean isDebugEnabled() {
-        return isLogLevelEnabled(LOG_LEVEL_DEBUG);
-    }
-
-    @Override
-    public boolean isDebugEnabled(Marker marker) {
-        return isDebugEnabled();
-    }
-
-    @Override
-    public boolean isInfoEnabled() {
-        return isLogLevelEnabled(LOG_LEVEL_INFO);
-    }
-
-    @Override
-    public boolean isInfoEnabled(Marker marker) {
-        return isInfoEnabled();
-    }
-
-    @Override
-    public boolean isWarnEnabled() {
-        return isLogLevelEnabled(LOG_LEVEL_WARN);
-    }
-
-    @Override
-    public boolean isWarnEnabled(Marker marker) {
-        return isWarnEnabled();
-    }
-
-    @Override
-    public boolean isErrorEnabled() {
-        return isLogLevelEnabled(LOG_LEVEL_ERROR);
-    }
-
-    @Override
-    public boolean isErrorEnabled(Marker marker) {
-        return isErrorEnabled();
-    }
-
-    private boolean isLogLevelEnabled(Level logLevel) {
-        return logLevel.toInt() >= (isJPlagLog() ? CURRENT_LOG_LEVEL.toInt() : LOG_LEVEL_FOR_EXTERNAL_LIBRARIES.toInt());
-    }
-
-    private boolean isJPlagLog() {
-        return this.name.startsWith("de.jplag.");
-    }
-
     private String computeShortName() {
         return name.substring(name.lastIndexOf(".") + 1);
-    }
-
-    @Override
-    protected String getFullyQualifiedCallerName() {
-        return null; //does not seem to be used by anything, but is required by SLF4J
     }
 }
